Remove hardcoded SIMD width from stim/circuit directory
This commit removes the dependecy of all Circuit components on
`MAX_BITWORD_WIDTH`. It does this via templatisation of
`decompose_mpp_operation` and some light reorganisation of files
to facilitate this.
